+++ api/python/quilt3/admin/types.py
@@ -50,3 +50,21 @@ class SSOConfig:
 class TabulatorTable:
     name: str
     config: str
+
+
+@pydantic.dataclasses.dataclass
+class Bucket:
+    name: str
+    title: str
+    icon_url: Optional[str]
+    description: Optional[str]
+    overview_url: Optional[str]
+    tags: Optional[List[str]]
+    relevance_score: int
+    last_indexed: Optional[datetime]
+    sns_notification_arn: Optional[str]
+    scanner_parallel_shards_depth: Optional[int]
+    skip_meta_data_indexing: Optional[bool]
+    file_extensions_to_index: Optional[List[str]]
+    index_content_bytes: Optional[int]
+    prefixes: List[str]