+++ catalog/app/utils/BucketPreferences/BucketPreferences.ts
@@ -14,6 +14,7 @@ import * as YAML from 'utils/yaml'
 export type ActionPreferences = Record<
   | 'copyPackage'
   | 'createPackage'
+  | 'deleteObject'
   | 'deleteRevision'
   | 'downloadObject'
   | 'downloadPackage'
@@ -143,11 +144,15 @@ const defaultGallery: GalleryPreferences = {
 }
 
 function getDefaultPreferences(bucket?: string): BucketPreferences {
+  // NOTE: if you add values here,
+  //       also update UI values for editor
+  //       in app/components/FileEditor/QuiltConfigEditor/BucketPreferences/State.tsx
   return {
     ui: {
       actions: {
         copyPackage: true,
         createPackage: true,
+        deleteObject: false,
         deleteRevision: false,
         downloadObject: true,
         downloadPackage: true,