+++ docs/CHANGELOG.md
@@ -16,6 +16,12 @@ Entries inside each section should be ordered by type:
 
 # Changelog
 
+## 7.1.0 - 2025-12-22
+
+### Python API
+
+* [Added] `quilt3.admin.buckets` sub-module for managing bucket configuration ([#4667](https://github.com/quiltdata/quilt/pull/4667))
+
 ## 7.0.0 - 2025-08-25
 
 ### Python API