+++ lambdas/shared/CHANGELOG.md
@@ -17,6 +17,9 @@ where verb is one of
 
 ## Changes
 
+- [Changed] Migrate to proper package structure ([#4652](https://github.com/quiltdata/quilt/pull/4652))
+- [Changed] Switch to uv ([#4652](https://github.com/quiltdata/quilt/pull/4652))
+- [Changed] Upgrade to Python 3.13 ([#4652](https://github.com/quiltdata/quilt/pull/4652))
 - [Changed] Upgrade to Python 3.11 ([#4241](https://github.com/quiltdata/quilt/pull/4241))
 - [Changed] Specify dependencies explicitly ([#4241](https://github.com/quiltdata/quilt/pull/4241))
 - [Added] Bootstrap the change log ([#4241](https://github.com/quiltdata/quilt/pull/4241))