+++ lambdas/tabular_preview/CHANGELOG.md
@@ -17,5 +17,8 @@ where verb is one of
 
 ## Changes
 
+- [Added] Preview h5ad (anndata) files ([#4636](https://github.com/quiltdata/quilt/pull/4636))
+- [Changed] Switch to uv ([#4654](https://github.com/quiltdata/quilt/pull/4654))
+- [Changed] Upgrade to Python 3.13 ([#4654](https://github.com/quiltdata/quilt/pull/4654))
 - [Changed] Upgrade to Python 3.11 ([#4241](https://github.com/quiltdata/quilt/pull/4241))
 - [Added] Bootstrap the change log ([#4241](https://github.com/quiltdata/quilt/pull/4241))